requirements.txt
Django==1.8.7
Pillow==3.1.0
django-cors-headers==1.1.0
